github: add some basic GH Actions to build the notebook
This patch adds GH Actions to built the HTML, PDF, and EPUB variants
of The SELinux Notebook when the GH repo is updated or a pull request
is submitted.  This should make it much easier for people to validate
their changes as well as providing fully rendered copies for each
revision moving forward.

Signed-off-by: Paul Moore <paul@paul-moore.com>