notebook: auditing: list noaudit exemptions
In the general case a rejected capability check will result in an audit
event.  There are however some instances in the kernel where denied
capability checks are not audited, which could lead to differences in
behavior between enforcing and permissive mode.

Document this fact and list (hopefully) all occurrences in kernel v6.4.

Signed-off-by: Christian Göttsche <cgzones@googlemail.com>